requirements.txt
flask
flask-sqlalchemy
flask-cors
python-dotenv
PyJWT
pymupdf
google-genai
psycopg2-binary
